[web] Merge landing pages into new one
