--- packages/bento-web/pages/dashboard.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/dashboard-landing/DashboardLandingPage';
