+++ packages/bento-web/pages/index.ts
@@ -1,1 +1,1 @@
-export { default, getStaticProps } from '@/landing-v1/LandingPage';
+export { default, getStaticProps } from '@/landing/LandingPage';
