--- packages/bento-web/pages/v1/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@/landing-v1/LandingPage';
