--- packages/bento-web/src/landing-v1/components/SectionBadge.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-import { Badge } from '@/components/system';
-
-export const SectionBadge = styled(Badge)`
-  @media screen and (max-width: 600px) {
-    font-size: 12px;
-  }
-`;
