--- packages/bento-web/src/landing-v1/components/SectionTitle.ts
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-export const SectionTitle = styled.h2`
-  margin-top: 27px;
-
-  font-weight: 900;
-  font-size: 52px;
-  line-height: 103%;
-  letter-spacing: 0.01em;
-  color: #ffffff;
-
-  @media screen and (max-width: 600px) {
-    font-size: 36px;
-  }
-`;
