--- packages/bento-web/src/landing-v2/LandingPage.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-import { HeaderSection } from './sections/HeaderSection';
-
-export const LandingPage = () => {
-  return (
-    <Container>
-      <HeaderSection />
-    </Container>
-  );
-};
-
-const Container = styled.div``;
