--- packages/bento-web/src/landing-v2/index.ts
@@ -1,1 +1,0 @@
-export { LandingPage } from './LandingPage';
