+++ packages/bento-web/src/landing/components/SectionBadge.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+import { Badge } from '@/components/system';
+
+export const SectionBadge = styled(Badge)`
+  @media screen and (max-width: 600px) {
+    font-size: 12px;
+  }
+`;
